Centralize detection types and tighten cleanup check
